+++ wasi_virt_layer-cli/src/commands/new.rs
@@ -163,6 +163,8 @@ export!(ComponentABI);
 mod process {
     use super::*;
     plug_process!(StandardProcess, anonymous, self);
+    plug_random!(StandardRandom, anonymous, self);
+    plug_sched!(DefaultSched, anonymous, self);
 }
 
 mod env {
@@ -257,6 +259,8 @@ plug_thread!(
 mod process {
     use super::*;
     plug_process!(StandardProcess, anonymous, self);
+    plug_random!(StandardRandom, anonymous, self);
+    plug_sched!(DefaultSched, anonymous, self);
 }
 
 mod env {